Add note about pip using cache
I did not see a way to open an issue and found this was a hiccup for me when installing multiple versions of mpi4py on my system. If the language needs some work I would be happy to rewrite but I think this note would be helpful for users.