GPU: Prevent exception chaining on getting CPU/GPU buffers
If an object does not provide Python's (CPU) buffer interface nor the CUDA
array interface (CAI), the exception traceback contains references to the
unavailability of the CAI, which would be quite confusing for regular users
not dealing with GPU arrays. This fix prevents the CAI lookup error from
showing up on exception tracebacks.