mpi4py.futures: Add `_max_workers` property to `MPIPoolExecutor`
This is a new undocumented and private API. Its only purpose is improving
compatibility with the `{Process|Thread}PoolExecutor` classes from
the `concurrent.futures` stdlib module.

Despite the `max` in its name, `executor._max_workers` return the exact
number of MPI worker processes the executor is using. This information may
help to decide how many tasks to dispatch to the executor, or how to better
batch them to minimize scheduling overhead and maximize performance.

WARNING: This API is private and provisional, it may change or go away at
any time without further notice or deprecation warning.